+++ NEWS.md
@@ -1,6 +1,7 @@
-# meteoR 0.0.9101
+# metR 0.0.9101
 - New function: GetTopography()
-- ReadNetCDF() now can output arrays and vectors. 
+- ReadNetCDF() now can output arrays and vectors.
+- Changed name. 
 
 # meteoR 0.0.9100
 - Full (not perfect) documentation.